chore: move results file logic from cache service
