+++ packages/backend/src/ee/index.ts
@@ -204,6 +204,8 @@ export async function getEnterpriseAppArguments(): Promise<EnterpriseAppArgument
                     queryHistoryModel: models.getQueryHistoryModel(),
                     cacheService: repository.getCacheService(),
                     savedSqlModel: models.getSavedSqlModel(),
+                    resultsFileModel: models.getResultsFileModel(),
+                    storageClient: clients.getResultsFileStorageClient(),
                 }),
             cacheService: ({ models, context, clients }) =>
                 new CommercialCacheService({
