+++ packages/backend/src/services/AsyncQueryService/AsyncQueryService.ts
@@ -21,6 +21,7 @@ import {
     type ExecuteAsyncQueryRequestParams,
     type ExecuteAsyncSavedChartRequestParams,
     type ExecuteAsyncUnderlyingDataRequestParams,
+    ExpiredError,
     Explore,
     ForbiddenError,
     formatRow,
@@ -41,6 +42,7 @@ import {
     isVizTableConfig,
     ItemsMap,
     MetricQuery,
+    NotFoundError,
     type Organization,
     PivotIndexColum,
     prefixPivotConfigurationReferences,
@@ -48,6 +50,7 @@ import {
     QueryExecutionContext,
     QueryHistoryStatus,
     type ResultColumns,
+    ResultRow,
     type RunQueryTags,
     SessionUser,
     SortBy,
@@ -58,13 +61,18 @@ import {
     WarehouseClient,
 } from '@lightdash/common';
 import { SshTunnel } from '@lightdash/warehouses';
+import { createInterface } from 'readline';
+import type { S3ResultsFileStorageClient } from '../../clients/ResultsFileStorageClients/S3ResultsFileStorageClient';
+import type { DbResultsCacheUpdate } from '../../database/entities/resultsFile';
 import { measureTime } from '../../logging/measureTime';
 import type { QueryHistoryModel } from '../../models/QueryHistoryModel';
+import { ResultsFileModel } from '../../models/ResultsFileModel/ResultsFileModel';
 import type { SavedSqlModel } from '../../models/SavedSqlModel';
 import { applyLimitToSqlQuery } from '../../queryBuilder';
 import { wrapSentryTransaction } from '../../utils';
 import type { ICacheService } from '../CacheService/ICacheService';
 import {
+    type CacheHitCacheResult,
     CreateCacheResult,
     MissCacheResult,
     ResultsCacheStatus,
@@ -95,6 +103,8 @@ type AsyncQueryServiceArguments = ProjectServiceArguments & {
     queryHistoryModel: QueryHistoryModel;
     cacheService?: ICacheService;
     savedSqlModel: SavedSqlModel;
+    resultsFileModel: ResultsFileModel;
+    storageClient: S3ResultsFileStorageClient;
 };
 
 export class AsyncQueryService extends ProjectService {
@@ -104,16 +114,24 @@ export class AsyncQueryService extends ProjectService {
 
     savedSqlModel: SavedSqlModel;
 
+    private readonly resultsFileModel: ResultsFileModel;
+
+    private readonly storageClient: S3ResultsFileStorageClient;
+
     constructor({
         queryHistoryModel,
         cacheService,
         savedSqlModel,
+        resultsFileModel,
+        storageClient,
         ...projectServiceArgs
     }: AsyncQueryServiceArguments) {
         super(projectServiceArgs);
         this.queryHistoryModel = queryHistoryModel;
         this.cacheService = cacheService;
         this.savedSqlModel = savedSqlModel;
+        this.resultsFileModel = resultsFileModel;
+        this.storageClient = storageClient;
     }
 
     // ! Duplicate of SavedSqlService.hasAccess
@@ -163,6 +181,164 @@ export class AsyncQueryService extends ProjectService {
             projectUuid: savedChart.project.projectUuid,
             organizationUuid: savedChart.organization.organizationUuid,
         });
+    }
+
+    private getCacheExpiresAt(baseDate: Date) {
+        return new Date(
+            baseDate.getTime() +
+                this.lightdashConfig.results.cacheStateTimeSeconds * 1000,
+        );
+    }
+
+    async createOrGetExistingCache(
+        projectUuid: string,
+        cacheIdentifiers: {
+            sql: string;
+            timezone?: string;
+        },
+        invalidateCache: boolean = false,
+    ): Promise<CreateCacheResult> {
+        // Generate cache key from project and query identifiers
+        const cacheKey = ResultsFileModel.getCacheKey(
+            projectUuid,
+            cacheIdentifiers,
+        );
+
+        // Check if cache already exists
+        const existingCache = await this.cacheService?.findCachedResultsFile(
+            projectUuid,
+            cacheIdentifiers,
+        );
+
+        // Case 1: Valid cache exists and not being invalidated
+        if (existingCache && !invalidateCache) {
+            return existingCache;
+        }
+
+        // Create upload stream for storing results
+        const { write, close } = this.storageClient.createUploadStream(
+            cacheKey,
+            DEFAULT_RESULTS_PAGE_SIZE,
+        );
+
+        const now = new Date();
+        const newExpiresAt = this.getCacheExpiresAt(now);
+
+        // Case 2: No valid cache exists - upsert cache entry
+        const createdCache = await this.resultsFileModel.create({
+            cache_key: cacheKey,
+            project_uuid: projectUuid,
+            expires_at: newExpiresAt,
+            total_row_count: null,
+            status: ResultsCacheStatus.PENDING,
+        });
+
+        if (!createdCache) {
+            await close();
+            throw new Error('Failed to create cache');
+        }
+
+        return {
+            cacheKey: createdCache.cache_key,
+            createdAt: createdCache.created_at,
+            updatedAt: createdCache.updated_at,
+            expiresAt: createdCache.expires_at,
+            write,
+            close,
+            cacheHit: false,
+            totalRowCount: null,
+        };
+    }
+
+    async getCachedResultsPage(
+        cacheKey: string,
+        projectUuid: string,
+        page: number,
+        pageSize: number,
+        formatter: (row: ResultRow) => ResultRow,
+    ) {
+        const cache = await this.resultsFileModel.find(cacheKey, projectUuid);
+
+        if (!cache) {
+            // TODO: throw a specific error the FE will respond to
+            throw new NotFoundError(
+                `Cache not found for key ${cacheKey} and project ${projectUuid}`,
+            );
+        }
+
+        if (cache.expires_at < new Date()) {
+            await this.resultsFileModel.delete(cacheKey, projectUuid);
+
+            // TODO: throw a specific error the FE will respond to
+            throw new ExpiredError(
+                `Cache expired for key ${cacheKey} and project ${projectUuid}`,
+            );
+        }
+
+        const cacheStream = await this.storageClient.getDowloadStream(cacheKey);
+
+        const rows: ResultRow[] = [];
+        const rl = createInterface({
+            input: cacheStream,
+            crlfDelay: Infinity,
+        });
+
+        const startLine = (page - 1) * pageSize;
+        const endLine = startLine + pageSize;
+        let nonEmptyLineCount = 0;
+
+        for await (const line of rl) {
+            if (line.trim()) {
+                if (
+                    nonEmptyLineCount >= startLine &&
+                    nonEmptyLineCount < endLine
+                ) {
+                    rows.push(formatter(JSON.parse(line)));
+                }
+                nonEmptyLineCount += 1;
+            }
+        }
+
+        return {
+            rows,
+            totalRowCount: cache.total_row_count ?? 0,
+            expiresAt: cache.expires_at,
+        };
+    }
+
+    async updateCache(
+        cacheKey: string,
+        projectUuid: string,
+        update: DbResultsCacheUpdate,
+    ) {
+        await this.resultsFileModel.update(cacheKey, projectUuid, update);
+    }
+
+    async deleteCache(cacheKey: string, projectUuid: string) {
+        await this.resultsFileModel.delete(cacheKey, projectUuid);
+    }
+
+    async findCache(
+        cacheKey: string,
+        projectUuid: string,
+    ): Promise<CacheHitCacheResult | undefined> {
+        const cache = await this.resultsFileModel.find(cacheKey, projectUuid);
+
+        if (cache) {
+            return {
+                cacheKey,
+                createdAt: cache.created_at,
+                updatedAt: cache.updated_at,
+                expiresAt: cache.expires_at,
+                cacheHit: true,
+                write: undefined,
+                close: undefined,
+                totalRowCount: cache.total_row_count ?? 0,
+                status: cache.status,
+            };
+        }
+
+        return undefined;
     }
 
     async cancelAsyncQuery({
@@ -286,10 +462,7 @@ export class AsyncQueryService extends ProjectService {
                 };
             case QueryHistoryStatus.PENDING:
                 if (resultsCacheEnabled && cacheKey && this.cacheService) {
-                    const cache = await this.cacheService.findCache(
-                        cacheKey,
-                        projectUuid,
-                    );
+                    const cache = await this.findCache(cacheKey, projectUuid);
 
                     if (cache?.status === ResultsCacheStatus.READY) {
                         // If the cache is ready, we update the query history status to READY and return the current state
@@ -338,7 +511,7 @@ export class AsyncQueryService extends ProjectService {
                 durationMs,
             } = await measureTime(
                 () =>
-                    cacheService.getCachedResultsPage(
+                    this.getCachedResultsPage(
                         cacheKey,
                         projectUuid,
                         page,
@@ -621,14 +794,10 @@ export class AsyncQueryService extends ProjectService {
             // Wait for the cache to be written before marking the query as ready
             if (resultsCache) {
                 await resultsCache.close();
-                await this.cacheService?.updateCache(
-                    resultsCache.cacheKey,
-                    projectUuid,
-                    {
-                        status: ResultsCacheStatus.READY,
-                        total_row_count: totalRows,
-                    },
-                );
+                await this.updateCache(resultsCache.cacheKey, projectUuid, {
+                    status: ResultsCacheStatus.READY,
+                    total_row_count: totalRows,
+                });
                 this.analytics.track({
                     userId: user.userUuid,
                     event: 'results_cache.write',
@@ -677,10 +846,7 @@ export class AsyncQueryService extends ProjectService {
 
             // When the query fails, we delete the cache entry
             if (resultsCache) {
-                await this.cacheService?.deleteCache(
-                    resultsCache.cacheKey,
-                    projectUuid,
-                );
+                await this.deleteCache(resultsCache.cacheKey, projectUuid);
                 this.analytics.track({
                     userId: user.userUuid,
                     event: 'results_cache.delete',
@@ -851,15 +1017,14 @@ export class AsyncQueryService extends ProjectService {
                     let resultsCache: CreateCacheResult | undefined;
 
                     if (resultsCacheEnabled && this.cacheService) {
-                        resultsCache =
-                            await this.cacheService.createOrGetExistingCache(
-                                projectUuid,
-                                {
-                                    sql: query,
-                                    timezone: metricQuery.timezone,
-                                },
-                                args.invalidateCache,
-                            );
+                        resultsCache = await this.createOrGetExistingCache(
+                            projectUuid,
+                            {
+                                sql: query,
+                                timezone: metricQuery.timezone,
+                            },
+                            args.invalidateCache,
+                        );
 
                         if (!resultsCache.cacheHit) {
                             this.analytics.track({
