+++ packages/backend/src/services/ServiceRepository.ts
@@ -508,6 +508,8 @@ export class ServiceRepository
                     userModel: this.models.getUserModel(),
                     queryHistoryModel: this.models.getQueryHistoryModel(),
                     savedSqlModel: this.models.getSavedSqlModel(),
+                    resultsFileModel: this.models.getResultsFileModel(),
+                    storageClient: this.clients.getResultsFileStorageClient(),
                 }),
         );
     }
